requirements.txt
appdirs==1.4.0
Babel==2.3.4
blinker==1.4
click==6.7
decorator==4.0.11
defusedxml==0.4.1
Flask==0.12
Flask-Admin==1.4.2
Flask-Babel==0.11.1
Flask-CacheControl==0.1.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
guess-language-spirit==0.5.2
itchat==1.2.22
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
olefile==0.44
packaging==16.8
pbr==1.10.0
peewee==2.8.5
Pillow==4.0.0
psutil==5.1.3
pyodbc==4.0.3
pyparsing==2.1.10
pypiwin32==219
python3-openid==3.0.10
pytz==2016.10
pywin32==220
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.15
Whoosh==2.7.4
WTForms==2.1
